requirements.txt
boto3>=1.34.0
botocore>=1.34.0
click>=8.1.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0